requirements.txt
pandas>=2.2.0
matplotlib>=3.8.0
langgraph>=0.2.0
google-genai>=0.1.0
pydantic>=2.7.0
fastparquet>=2024.11.0
python-dotenv>=1.1.1
ipykernel>=6.30.1